rooms: merge NO_ROOM definitions
This merges the two NO_ROOM concepts so -1 is the standard. Legacy
sector pit and sky portals will be converted on level load. This
ultimately paves the way for allowing more than 256 rooms in a level
via alternate sector data provided by injections (to be implemented).

+++ src/libtrx/include/libtrx/game/rooms/types.h
@@ -46,8 +46,8 @@ typedef struct {
 #endif
     TRIGGER *trigger;
     struct {
-        uint8_t pit;
-        uint8_t sky;
+        int16_t pit;
+        int16_t sky;
         int16_t wall;
     } portal_room;
     struct {
